% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{arraycorr_gen}
\alias{arraycorr_gen}
\title{Generate array of fisher-z transformed correlations from raw csv correlations}
\usage{
arraycorr_gen(corrloc, freqs)
}
\arguments{
\item{corrloc}{location of rds correlation file, typically generated from create.df or gencorr}

\item{freqs}{frequencies that are desired to have in the array}
}
\value{
a list with 1) 3-way array of correlations, with frequencies in rows, images, in columns, and
repetitions in the third dimension, 2) the matching of the correlations and 3) the knife pair for the correlation
}
\description{
This function reads the raw correlations csv files and returns an array with p x q x n , where p is the frequencies, q is the images, and n is the sample size
}
\details{
Given a csv file for correlations (corrloc) and a set of frequencies (freq) will return the 3-way array
 (corrs), where ith slice (indicated in the 3d array mode) contains the ith matrix feature
}
\author{
Carlos Llosa

Geoffrey Z. Thompson
}
