% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{classif_gen}
\alias{classif_gen}
\title{classify a set of correlations based on a training and a testing path}
\usage{
classif_gen(
  corrloc_train,
  corrloc_test,
  prob_loc = NULL,
  boxplot_loc = NULL,
  freqs = c(5, 10, 20),
  df = 5,
  verb = T
)
}
\arguments{
\item{corrloc_train}{directory of the training set}

\item{corrloc_test}{directory of the testing set (to classify)}

\item{prob_loc}{location where one wants to save the probabilities (includying file name)}

\item{boxplot_loc}{location where one wants to save the probabilities boxplots (includying file name)}

\item{freqs}{frequencies to be used in the training/testing}

\item{df}{degrees of freedom in the matrix-t distribution that will be used}

\item{verb}{T or F indicating whether one wants to see progress}
}
\value{
a list with 1) posterior probabilities of match 2) the boxplot 3) parameters used in the training and 4) ML fits
}
\description{
this function will take the path of the training and testing sets, and classify
}
\author{
Carlos Llosa-Vite

Geoffrey Z. Thompson
}
