% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{classify}
\alias{classify}
\title{classify a set of correlations based on a training and a testing path}
\usage{
classify(train, test, freqs = c(5, 10, 20), df = 5)
}
\arguments{
\item{train}{data frame to use as testing set}

\item{test}{data frame to use as training set}

\item{freqs}{frequencies to be used in the training/testing}

\item{df}{degrees of freedom in the matrix-t distribution that will be used}

\item{verb}{T or F indicating whether one wants to see progress}
}
\value{
a list with 1) posterior probabilities of match 2) parameters used in
the training and 3) ML fits
}
\description{
this function will take the path of the training and testing sets, and
classify
}
\details{
this is a copy of classif_gen(), but rewritten to take in data frames, 
and excludes any saving, plotting, etc
}
\author{
Josh Berlinski

Carlos Llosa-Vite

Geoffrey Z. Thompson
}
