% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{cor2d}
\alias{cor2d}
\title{Correlation Between Two Array Slices}
\usage{
cor2d(mats, ind.mat)
}
\arguments{
\item{mats}{An array of dimension (m1 x m2 x 2), where m1 x m2 is the image size which contains the base and tip}

\item{ind.mat}{A matrix of size (m1 x m2) that indicates the regions where to compute the correlations betwee base and tip.
Typically generated by freq.map.radial().}
}
\value{
a correlation (between 0 and 1)
}
\description{
this function finds the correlation between two matrices along pre-specified locations
}
\author{
Carlos Llosa-Vite, \email{llosacarlos2@gmail.com}

Geoffrey Z. Thompson
}
