% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastread.R
\name{fread_fft_single}
\alias{fread_fft_single}
\title{Read data quickly, from CSV files where there is no Base/Tip correpsondence}
\usage{
fread_fft_single(
  img_no,
  sample_names,
  operator,
  fdir,
  file_id = "_FFT",
  dim = c(512, 1023),
  prepend_zero = TRUE,
  verbose = TRUE,
  stopifmissing = TRUE,
  skip = 4
)
}
\arguments{
\item{img_no}{numeric; number of images per surface}

\item{sample_names}{character vector; each element is a string corresponding to the fractured surface being
investigated}

\item{operator}{character; who imaged the sample, used to construct the end of the file to be read, with `file_id`}

\item{fdir}{character; director which holds the csv files}

\item{file_id}{the end of the filename, typically "_FFT", which will be used to form the filename '.*operator_FFT.csv'}

\item{prepend_zero}{logical; for images 1-9, does the filename format as 01-09?}

\item{verbose}{logical; should progress be printed?}

\item{stopifmissing}{logical; should the program be halted if a file is missing?}

\item{skip}{numeric; how many lines at the start correspond to metadata? this is usually 4 for the magnification and
resolution. The resolution should be the last line (ie, line number `skip`)}

\item{dim;}{numeric vector; size of the FFT matrix}
}
\description{
The images should now exist in a 512 by 1023 (or other) matrix that is stored in a CSV file, with one file per image.
This function will loop through files in a specified directory and either:
  1) stores the 6 image pairs for each surface in a list which is returned
  2) creates the list in 1), and computes a data frame of correlation which is returned
}
\details{
This function should be used over `fread_fft` when there are no Base/Tip
pairs. For example, for our ceramic cup, there are surfaces CU01, CU02, ...
where C01 and C02 are matching surfaces. In this case, there is no need for
the "surface_names" argument, and the dimension of the output should be
reduced. As a consequence to this, there also needs to be a new create_df
function for this kind of output.
}
\author{
josh berlinski
}
