% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{freq.map.radial}
\alias{freq.map.radial}
\title{Create regions where to find correlations based on frequency bands, image size, and magnification}
\usage{
freq.map.radial(freqs, freq_mult, image.dimensions)
}
\arguments{
\item{freqs}{This is an alternative to specifying cuts (note that you should only specify one or the other). Suppose that you wished to calculate
correlations for the 5-10 and 20-30 frequency ranges, but not 10-20. Then you could use this argument by passing a matrix with two columns and a
 number of rows equal to the number of frequency ranges you want to perform calculations for. The first column indicates the lower number in the
 frequency range, and the second column specifies the upper number in the frequency range.}

\item{freq_mult}{This argument is the multiplier that tells you the at a given row or column index. This changes depending on the image
magnification.}

\item{image.dimensions}{Image dimensions (not neccesarity 512 x 1023)}
}
\value{
matrix of indicator variables
}
\description{
This function creates a matrix of indicator variables specifying which cells in the 512 by 1023 matrix should be used in the calculation of the
correlation for a given frequency range. I would suggest writing an entire matrix from this function to a csv file to get a handle on what it's
doing.
}
\examples{
freqs <- freq.ranges(cuts = c(5,10,20))
ind.mats <- freq.map.radial(freqs = freqs, freq_mult = 1.775, image.dimensions = c(512,1023))
dim(ind.mats)
ind.mats[c(502:512), c(506:518),1]
}
\author{
Geoffrey Thompson

Geoffrey Z. Thompson
}
