% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{gencorr}
\alias{gencorr}
\title{Generate correlations from raw FFT data}
\usage{
gencorr(
  img.no,
  sample.names,
  base.names,
  tip.names = base.names,
  base.directory,
  tip.directory = base.directory,
  surface = c("Base", "Tip"),
  verb = T,
  stopmiss = T,
  cuts = NA,
  freqs,
  freq_mult = 1.775,
  map_func = freq.map.radial,
  corr_csvloc = NULL,
  corr_pdfloc = NULL
)
}
\arguments{
\item{img.no}{number of (usually) overlapping images}

\item{sample.names}{This is another character vector where each element is a string corresponding to the knives we wish to include.
In the image.names example, the element of this vector corresponding to knife T01 would be T01.}

\item{base.names}{names of the base images. Usually if operator A did it is A_FFT}

\item{tip.names}{usually the same name as the base, but might change.}

\item{base.directory}{The file path for where the bases are located}

\item{tip.directory}{typically the same as base.directory}

\item{surface}{Typically  c("Base", "Tip"), but this might change. For instance they might be  c("base", "tip") instead}

\item{verb}{show progress? True or False}

\item{stopmiss}{should the function stop when there is missing data?}

\item{cuts}{numeric vector of cut points for specifying frequency ranges over which we would like to calculate correlations.
For example, if 5-10 and 10-20 frequency ranges are desired, set cuts as c(5,10,20).}

\item{freqs}{This is an alternative to specifying cuts (note that you should only specify one or the other). Suppose that you wished to calculate
correlations for the 5-10 and 20-30 frequency ranges, but not 10-20. Then you could use this argument by passing a matrix with two columns and a
 number of rows equal to the number of frequency ranges you want to perform calculations for. The first column indicates the lower number in the
 frequency range, and the second column specifies the upper number in the frequency range.}

\item{freq_mult}{This argument is the multiplier that tells you the at a given row or column index. This changes depending on the image
magnification.}

\item{map_func}{This argument is a function that takes in three arguments: freqs, freq\_mult, and image.dimensions, and produces a 3-D array
where, along the third dimension, each element is a matrix of zeros and ones specifying the cells in a given image that are relevant for
calculations at a given frequency range. Each of these matrices will be passed to func to specify the cells from a specific frequency range.
The default function used here will likely not need to change.}

\item{corr_csvloc}{location for csv files of correlations}

\item{corr_pdfloc}{location for pdf file of correlation histograms}
}
\value{
correlations along all images, pairs and frequencies, will also save pdf and csv correlation files.
}
\description{
This function reads the data from myread.data, saves it with create.df and creates histograms with hist_pdf
}
\author{
Carlos Llosa

Geoffrey Z. Thompson
}
