% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{hist_pdf}
\alias{hist_pdf}
\title{Create correlation histograms}
\usage{
hist_pdf(cordf, filename, bns, data.name)
}
\arguments{
\item{cordf}{data frame of correlations, as created by create.df}

\item{filename}{Name of pdf file to be saved, includying directory}

\item{bns}{histogram bins}

\item{data.name}{name of data, will appear in the histograms}
}
\value{
won't return anything, will just save the pdf
}
\description{
This function will save a pdf file of correlation histograms, with one frequency band per page
}
\author{
Geoffrey Thompson
}
