% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{reshape_correlation}
\alias{reshape_correlation}
\title{Generate array of Fisher-Z transformed correlations from data frame of
correlations}
\usage{
reshape_correlation(correlation_df, frequencies)
}
\arguments{
\item{correlation_df}{data frame containing correlation scores, generated
from create.df}

\item{frequencies}{frequencies that are desired to have in the array}
}
\value{
a list with 1) 3-way array of correlations with frequencies in rows,
images in columns, and repetitions in the third dimension 2) the matching of
the correlations and 3) the knife pair for the correlation
}
\description{
This function reads the raw correlations csv files and returns an array with
p x q x n , where p is the frequencies, q is the images, and n is the sample
size
}
\details{
This function is a copy of arraycor_gen, but takes in a data frame rather
than a csv path.

A caveat is that you need to read in the file with read.csv(), since this
depends on the column names being of the form X5.10, etc
}
\author{
Josh Berlinski

Carlos Llosa

Geoffrey Z. Thompson
}
